requirements.txt
bumpversion==0.5.3
docopt==0.6.2
json2yaml==1.1.1
jsonschema==2.6.0
py==1.4.34
pyaml==17.10.0
pytest==3.2.3
pytest-sugar==0.9.0
PyYAML==3.12
termcolor==1.1.0
